+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add ability to set sticky columns
   - Refactor table components
   - Add export to excel functionality
+  - Add power selects for editing business object entities
 
 ## [2.10.2] - Started: 6-30-20